+++ src/main/java/addressmode/Abx.java
@@ -3,7 +3,7 @@ package addressmode;
 public class Abx extends AddressingMode {
 
     @Override
-    public byte set() {
+    public short set() {
         return 0;
     }
 }
