+++ src/main/java/addressmode/Rel.java
@@ -3,7 +3,7 @@ package addressmode;
 public class Rel extends AddressingMode {
 
     @Override
-    public byte set() {
+    public short set() {
         return 0;
     }
 }
