+++ src/main/java/opcodes/Sec.java
@@ -2,7 +2,7 @@ package opcodes;
 
 public class Sec extends Opcode {
     @Override
-    public byte operate() {
+    public short operate() {
         return 0;
     }
 }
