+++ src/main/java/opcodes/Txs.java
@@ -2,7 +2,7 @@ package opcodes;
 
 public class Txs extends Opcode {
     @Override
-    public byte operate() {
+    public short operate() {
         return 0;
     }
 }
